test: Cypress 커스텀 함수 추가
