+++ cypress/support/commands.ts
@@ -1,37 +1,65 @@
-/// <reference types="cypress" />
-// ***********************************************
-// This example commands.ts shows you how to
-// create various custom commands and overwrite
-// existing commands.
-//
-// For more comprehensive examples of custom
-// commands please read more here:
-// https://on.cypress.io/custom-commands
-// ***********************************************
-//
-//
-// -- This is a parent command --
-// Cypress.Commands.add('login', (email, password) => { ... })
-//
-//
-// -- This is a child command --
-// Cypress.Commands.add('drag', { prevSubject: 'element'}, (subject, options) => { ... })
-//
-//
-// -- This is a dual command --
-// Cypress.Commands.add('dismiss', { prevSubject: 'optional'}, (subject, options) => { ... })
-//
-//
-// -- This will overwrite an existing command --
-// Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
-//
-// declare global {
-//   namespace Cypress {
-//     interface Chainable {
-//       login(email: string, password: string): Chainable<void>
-//       drag(subject: string, options?: Partial<TypeOptions>): Chainable<Element>
-//       dismiss(subject: string, options?: Partial<TypeOptions>): Chainable<Element>
-//       visit(originalFn: CommandOriginalFn, url: string, options: Partial<VisitOptions>): Chainable<Element>
-//     }
-//   }
-// }
+type TestRestaurantInfoType = {
+  imageSrc: string;
+  name: string;
+  distanceInMinutes: string;
+  description: string;
+  favoriteSrc: string;
+};
+
+type TestRestaurantFormInfoType = {
+  category: string;
+  name: string;
+  distanceInMinutes: string;
+  description: string;
+  link: string;
+};
+
+Cypress.Commands.add('openAddModal', () => {
+  cy.get('.modal.modal--open').should('not.exist');
+  cy.get('#header-add-button').click();
+  cy.get('.modal.modal--open').should('exist');
+});
+
+Cypress.Commands.add('closeAddModal', () => {
+  cy.get('.modal.modal--open').should('exist');
+  cy.get('#add-modal-cancel').click();
+});
+
+Cypress.Commands.add(
+  'writeRestaurantAddForm',
+  (input: TestRestaurantFormInfoType) => {
+    cy.get('#category').select(input.category);
+    if (input.name !== '') {
+      cy.get('#name').type(input.name);
+    }
+    cy.get('#distance').select(input.distanceInMinutes);
+    if (input.description !== '') {
+      cy.get('#description').type(input.description);
+    }
+    if (input.link !== '') {
+      cy.get('#link').type(input.link);
+    }
+    cy.get('#add-modal-submit').click();
+  }
+);
+
+Cypress.Commands.add(
+  'checkRestaurantItem',
+  (index: number, info: TestRestaurantInfoType) => {
+    cy.get('#restaurant-list-root .restaurant')
+      .eq(index)
+      .within(() => {
+        cy.get('.category-icon').should('have.attr', 'src', info.imageSrc);
+        cy.get('.restaurant__name').should('have.text', info.name);
+        cy.get('.restaurant__distance').should(
+          'have.text',
+          `캠퍼스부터 ${info.distanceInMinutes}분 내`
+        );
+        cy.get('.restaurant__description').should(
+          'have.text',
+          info.description
+        );
+        cy.get('.favorite-icon').should('have.attr', 'src', info.favoriteSrc);
+      });
+  }
+);
